% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{runVol2bird}
\alias{runVol2bird}
\title{convert pvol file}
\usage{
runVol2bird(inn, out)
}
\arguments{
\item{inn}{the input polar volume file as a string}

\item{the}{output h5 file for the vp}
}
\value{
A character with the vol to bird version
}
\description{
convert pvol file
}
\examples{
runVol2bird(system.file("extdata", "volume.h5", package = "bioRad"), t<-tempfile() )
vp<-bioRad::read_vpfiles(t)
plot(vp)

}
